requirements.txt
django==3.2
flake8==4.0
python-dotenv==1.0.0
django-split-settings==1.2.0
django-debug-toolbar==3.4.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
